chore: refactor configuration options
Refactor configuration options into storage, event-bus and index modules to ensure consistent parameter naming and configuration of these modules.

+++ event-bus/src/sqs.rs
@@ -4,14 +4,12 @@ use aws_sdk_sqs::{
 };
 use thiserror::Error;
 
-use crate::{Event, EventBus, EventConsumer};
+use crate::Event;
 
 #[derive(Debug, Error)]
 pub enum SqsError {
     #[error("Error from SQS: {0}")]
     Sqs(SqsSdkError),
-    #[error("No message received")]
-    NoMessage,
 }
 
 impl From<SqsSdkError> for SqsError {
@@ -26,7 +24,7 @@ pub struct SqsEventBus {
 }
 
 impl SqsEventBus {
-    pub async fn new() -> Result<Self, anyhow::Error> {
+    pub(crate) async fn new() -> Result<Self, anyhow::Error> {
         let region_provider = RegionProviderChain::default_provider().or_else(Region::new("eu-west-1"));
         let config = aws_config::from_env().region(region_provider).load().await;
         let client = Client::new(&config);
@@ -34,26 +32,22 @@ impl SqsEventBus {
     }
 }
 
-#[async_trait::async_trait]
-impl EventBus for SqsEventBus {
-    type Consumer<'m> = SqsConsumer<'m>;
-
-    async fn create(&self, topics: &[&str]) -> Result<(), anyhow::Error> {
+impl SqsEventBus {
+    pub(crate) async fn create(&self, topics: &[&str]) -> Result<(), anyhow::Error> {
         for topic in topics.iter() {
             self.client.create_queue().queue_name(topic.to_string()).send().await?;
         }
         Ok(())
     }
 
-    async fn subscribe(&self, _group: &str, topics: &[&str]) -> Result<Self::Consumer<'_>, anyhow::Error> {
+    pub(crate) async fn subscribe(&self, _group: &str, topics: &[&str]) -> Result<SqsConsumer, anyhow::Error> {
         Ok(SqsConsumer {
             client: &self.client,
             queues: topics.iter().map(|s| s.to_string()).collect(),
         })
     }
 
-    async fn send(&self, topic: &str, data: &[u8]) -> Result<(), anyhow::Error> {
-        // TODO
+    pub(crate) async fn send(&self, topic: &str, data: &[u8]) -> Result<(), anyhow::Error> {
         let s = core::str::from_utf8(data).unwrap();
         self.client
             .send_message()
@@ -70,10 +64,8 @@ pub struct SqsConsumer<'m> {
     queues: Vec<String>,
 }
 
-#[async_trait::async_trait]
-impl<'d> EventConsumer for SqsConsumer<'d> {
-    type Event<'m> = SqsEvent<'m> where Self: 'm;
-    async fn next<'m>(&'m self) -> Result<Option<Self::Event<'m>>, anyhow::Error> {
+impl<'d> SqsConsumer<'d> {
+    pub(crate) async fn next<'m>(&'m self) -> Result<Option<SqsEvent<'m>>, anyhow::Error> {
         let queue_futs: Vec<_> = self
             .queues
             .iter()
@@ -103,14 +95,16 @@ impl<'d> EventConsumer for SqsConsumer<'d> {
         Ok(None)
     }
 
-    async fn commit<'m>(&'m self, events: &[Self::Event<'m>]) -> Result<(), anyhow::Error> {
+    pub(crate) async fn commit<'m>(&'m self, events: &[Event<'m>]) -> Result<(), anyhow::Error> {
         for event in events {
-            self.client
-                .delete_message()
-                .queue_url(event.queue)
-                .set_receipt_handle(event.message.receipt_handle().map(|s| s.into()))
-                .send()
-                .await?;
+            if let Event::Sqs(event) = event {
+                self.client
+                    .delete_message()
+                    .queue_url(event.queue)
+                    .set_receipt_handle(event.message.receipt_handle().map(|s| s.into()))
+                    .send()
+                    .await?;
+            }
         }
         Ok(())
     }
@@ -121,13 +115,8 @@ pub struct SqsEvent<'m> {
     message: Message,
 }
 
-#[async_trait::async_trait]
-impl<'m> Event for SqsEvent<'m> {
-    fn payload(&self) -> Option<&[u8]> {
-        return self.message.body().map(|s| s.as_bytes());
-    }
-
-    fn topic(&self) -> &str {
-        self.queue
+impl<'m> SqsEvent<'m> {
+    pub(crate) fn payload(&self) -> Option<&[u8]> {
+        self.message.body().map(|m| m.as_bytes())
     }
 }
